# openapi.yaml
openapi: 3.0.1
info:
  title: E-Commerce(REST-API)
  version: 1.0.0
  description: A backend project for a E-Commerce site.
paths:
...
  /products/{id}:
    summary: To update name or specifications of a particular product.
    put:
      summary: To update name or specifications of a particular product.
      description: >-
        To update name or specifications of a particular product in the products
        table in the database schema.
      operationId: update_products
      parameters:
      - name: id
        in: path
        description: 'ID of the products'
        required: true
        schema:
          type: string
      responses:
        default:
          description: Default error sample response
          content:
            application/json:
              examples:
                PUT http://localhost:8080/products/112:
                  value: |-
                    {
                        "specifications":"Empty nothing written"
                    }

                    Update Completed
      tags:
        - products
    delete:
      summary: 'To delete a product in the products table. '
      description: To delete a product in the products table of the database schema.
      operationId: delete_products
      parameters:
      - name: id
        in: path
        description: 'ID of the products'
        required: true
        schema:
          type: string
      responses:
        default:
          description: Default error sample response
          content:
            application/json:
              examples:
                DELETE http://localhost:8080/products/112:
                  value: Deleted
      tags:
        - products